requirements.txt
fpdf
Pillow
natsort
PyPDF2

selenium
webdriver_manager
requests
pyyaml

